+++ BussinesLayer/Facades/CustomerFacade.cs
@@ -7,6 +7,8 @@ using BussinesLayer.Queries;
 using Riganti.Utils.Infrastructure.Core;
 using DataAccessLayer.Entities;
 using BussinesLayer.Filters;
+using System;
+using System.Data.Entity.Core;
 
 namespace BussinesLayer.Facades
 {
@@ -25,9 +27,15 @@ namespace BussinesLayer.Facades
 
         public void CreateCustomer(CustomerDTO customer)
         {
+            if (customer == null)
+                throw new ArgumentNullException("customer");
+
             using (var uow = UnitOfWorkProvider.Create())
             {
                 var created = Mapper.Map<Customer>(customer);
+                if (created == null)
+                    return;
+
                 CustomerRepository.Insert(created);
                 uow.Commit();
             }
@@ -37,42 +45,43 @@ namespace BussinesLayer.Facades
         {
             using (UnitOfWorkProvider.Create())
             {
-                var customer = CustomerRepository
-                    .GetById(customerId);
+                var customer = CustomerRepository.GetById(customerId);
+                if (customer == null)
+                    return null;
+
                 return Mapper.Map<CustomerDTO>(customer);
             }
         }
 
         public void UpdateCustomer(CustomerDTO customer)
         {
+            if (customer == null)
+                throw new ArgumentNullException("customer");
+
             using (var uow = UnitOfWorkProvider.Create())
             {
                 var retrieved = CustomerRepository.GetById(customer.Id);
+                if (retrieved == null)
+                    throw new ObjectNotFoundException("Customer hasn't been found");
+
                 Mapper.Map(customer, retrieved);
                 CustomerRepository.Update(retrieved);
                 uow.Commit();
             }
         }
 
-        public void DeleteCustomer(CustomerDTO Customer)
+        public void DeleteCustomer(CustomerDTO customer)
         {
-            using (var uow = UnitOfWorkProvider.Create())
-            {
-                var deleted = CustomerRepository.GetById(Customer.Id);
-                CustomerRepository.Delete(deleted);
-                uow.Commit();
-            }
-        }
+            if (customer == null)
+                throw new ArgumentNullException("customer");
 
-        public void DeleteCustomer(IEnumerable<CustomerDTO> customers)
-        {
             using (var uow = UnitOfWorkProvider.Create())
             {
-                foreach (var c in customers)
-                {
-                    var deleted = CustomerRepository.GetById(c.Id);
-                    CustomerRepository.Delete(deleted);
-                }
+                var deleted = CustomerRepository.GetById(customer.Id);
+                if (deleted == null)
+                    throw new ObjectNotFoundException("Customer hasn't been found");
+
+                CustomerRepository.Delete(deleted);
                 uow.Commit();
             }
         }
